django/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
cssselect2==0.6.0
Django==3.2.13 
#django-admin-volt==0.0.5
django-appconf==1.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filer==2.2.1
django-filter==2.1.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-polymorphic==3.1.0
django-select2==7.4.2
django-tinymce==3.4.0
djangorestframework==3.13.1
easy-thumbnails==2.8
environs==7.3.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
lxml==4.9.0
Markdown==3.3.4
marshmallow==3.17.0
martor==1.6.13
packaging==21.3
Pillow==9.1.1
promise==2.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dotenv==0.20.0
python-environ==0.4.54
pytz==2022.1
reportlab==3.6.10
requests==2.28.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
text-unidecode==1.3
tinycss2==1.1.1
Unidecode==1.1.2
urllib3==1.26.9
webencodings==0.5.1
Pillow==9.1.1
